+++ src/main/java/by/epamtc/digapply/service/impl/FacultyServiceImpl.java
@@ -8,8 +8,8 @@ import by.epamtc.digapply.entity.Faculty;
 import by.epamtc.digapply.entity.Subject;
 import by.epamtc.digapply.service.FacultyService;
 import by.epamtc.digapply.service.ServiceException;
-import by.epamtc.digapply.validator.EntityValidator;
-import by.epamtc.digapply.validator.EntityValidatorFactory;
+import by.epamtc.digapply.service.validation.EntityValidator;
+import by.epamtc.digapply.service.validation.EntityValidatorFactory;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
