+++ src/main/java/by/epamtc/digapply/service/impl/UserServiceImpl.java
@@ -7,8 +7,8 @@ import by.epamtc.digapply.entity.Role;
 import by.epamtc.digapply.entity.User;
 import by.epamtc.digapply.service.ServiceException;
 import by.epamtc.digapply.service.UserService;
-import by.epamtc.digapply.validator.EntityValidator;
-import by.epamtc.digapply.validator.EntityValidatorFactory;
+import by.epamtc.digapply.service.validation.EntityValidator;
+import by.epamtc.digapply.service.validation.EntityValidatorFactory;
 import org.apache.commons.codec.digest.DigestUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
